Align tokens on the last line vertically with the text field.
If the text field is larger than the JSTokenButton images, the
buttons may appear to float higher than the text, which looks
odd and distracting.

Note that this means that when the text field wraps to a new line,
the previous line's tokens may shift vertically to fit more compactly
with each other, if the text field is taller than the tokens.