Don't animate the frame around if the view isn't on screen yet.
This was causing a problem when the token field was in a view that
was presented modally. `layoutSubviews` was called after the view
was added to the window, but before it had any original frame. Thus,
the animateWithDuration call here was causing it to animate from
CGRectZero, which looked odd.